targhe_auto-py3.8/requirements.txt
# Requisiti minimi compatibili con Python 3.8
# Basati sulle versioni stabili supportate da Py3.8

ultralytics==8.2.0
fast-plate-ocr==1.1.0
opencv-python==4.8.0.74
opencv-python-headless==4.8.0.74
python-telegram-bot==20.8
python-dotenv==1.0.0
requests==2.31.0
numpy==1.24.4
torch==2.1.0
torchvision==0.16.0
Pillow==10.2.0
PyYAML==6.0.1
psutil==5.9.8
typing-extensions==4.9.0
